[CL-2525] Add button to remove cropped image
