+++ front/app/components/UI/RemoveImageButton/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  a11y_removeImage: {
+    id: 'app.components.UI.RemoveImageButton.a11y_removeImage',
+    defaultMessage: 'Remove',
+  },
+});
